+++ packages/cli/src/index.ts
@@ -1,11 +1,13 @@
 #!/usr/bin/env node
 import { Command } from 'commander'
-import temp from '@zdcode/template'
+import temp from './modules/template'
+import frame from './modules/frame'
 import packageJson from '../package.json'
 
 const program = new Command()
 
 temp(program)
+frame(program)
 
 program.version(
   packageJson.version,