+++ packages/cli/src/modules/frame/type.ts
@@ -1,0 +1,13 @@
+export type FrameType = {
+  id: number
+  name: string
+  root: string
+}
+export type Frame = {
+  id: number
+  name: string
+  git_url: string
+  type: number
+  branch: number
+  whether_init_git: boolean
+}